branch: support for shortcuts like @{-1} completed
branch command with options "edit-description", "set-upstream-to"
and "unset-upstream" expects a branch name.  Since ae5a6c3684
(checkout: implement "@{-N}" shortcut name for N-th last branch,
2009-01-17) a branch can be specified using shortcuts like @{-1}.
Those shortcuts need to be resolved when considering the
arguments.

Usage examples:

We can modify the description of the previously checked out branch
with:

$ git branch --edit--description @{-1}

We can modify the upstream of the previously checked out branch
with:

$ git branch --set-upstream-to upstream @{-1}
$ git branch --unset-upstream @{-1}

Signed-off-by: Rubén Justo <rjusto@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>